+++ cogs/reaction_roles.py
@@ -1,35 +1,43 @@
 import discord
 from discord.ext import commands
 from struc import db
+from discord.commands import (
+    slash_command,
+    Option
+)
 
 class ReactionRoles(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
 
-    @commands.command(help="Makes a reaction add a role to the user")
-    @commands.has_permissions(administrator=True)
-    async def set_reaction_role(self, ctx, message_id: int, role: discord.Role, emoji: str):
+    @slash_command()
+    async def set_reaction_role(self, ctx: discord.ApplicationContext, message_id: Option(int, "The message's ID"), role: Option(discord.Role, "The role to be given on reaction"), emoji: Option(str, "The emoji that is to be reacted with")):
+        if not ctx.author.guild_permissions.administrator:
+            raise commands.MissingPermissions(["Administrator"])
+
         exists = db.fetch("SELECT * FROM reaction_roles WHERE message_id = %s AND emoji = %s;", (message_id, emoji))
 
         if exists is not None:
-            await ctx.reply("This message already was set to a reaction role.", mention_author=False)
+            await ctx.respond("This message already was set to a reaction role.")
             return
 
         message = await ctx.fetch_message(message_id)
 
         if message is None:
-            await ctx.reply("Please provide an actual message ID.", mention_author=False)
+            await ctx.respond("Please provide an actual message ID.")
             return
 
         db.modify("INSERT INTO reaction_roles(guild_id, message_id, role_id, emoji) VALUES(%s,%s,%s,%s);", (ctx.guild.id, message.id, role.id, emoji))
 
         await message.add_reaction(emoji)
 
-        await ctx.reply("The reaction role has been set up.", mention_author=False)
+        await ctx.respond("The reaction role has been set up.")
 
-    @commands.command(help="Lists all reaction roles in the guild.")
-    @commands.has_permissions(administrator=True)
-    async def list_reaction_roles(self, ctx):
+    @slash_command()
+    async def list_reaction_roles(self, ctx: discord.ApplicationContext):
+        if not ctx.author.guild_permissions.administrator:
+            raise commands.MissingPermissions(["Administrator"])
+
         fetch = db.fetchall("SELECT * FROM reaction_roles WHERE guild_id = %s;", (ctx.guild.id,))
 
         fetch_embed = discord.Embed(title=f"Reaction roles in {ctx.guild.name}", color=discord.Color(0xF37F7F), description="")
@@ -40,15 +48,17 @@ class ReactionRoles(commands.Cog):
         if fetch is None or fetch == () or fetch == []:
             fetch_embed.description = "No reaction roles have been added."
 
-        await ctx.reply(embed=fetch_embed, mention_author=False)
+        await ctx.respond(embed=fetch_embed)
 
-    @commands.command(help="Remove a reaction role.")
-    @commands.has_permissions(administrator=True)
-    async def remove_reaction_role(self, ctx, message_id: int, emoji: str):
+    @slash_command()
+    async def remove_reaction_role(self, ctx: discord.ApplicationContext, message_id: Option(int, "The message's ID"), emoji: Option(str, "The emoji reaction which is to be removed")):
+        if not ctx.author.guild_permissions.administrator:
+            raise commands.MissingPermissions(["Administrator"])
+
         exists = db.fetch("SELECT * FROM reaction_roles WHERE message_id = %s AND emoji = %s;", (message_id, emoji))
 
         if exists is None:
-            await ctx.reply("That reaction role does not exist!")
+            await ctx.respond("That reaction role does not exist!")
             return
 
         db.modify("DELETE FROM reaction_roles WHERE message_id = %s AND emoji = %s;", (message_id, emoji))
@@ -58,7 +68,7 @@ class ReactionRoles(commands.Cog):
         if msg is not None:
             await msg.remove_reaction(emoji, self.bot.user)
 
-        await ctx.reply(f"Successfully deleted the specified reaction role.")
+        await ctx.respond(f"Successfully deleted the specified reaction role.")
 
     @commands.Cog.listener()
     async def on_raw_reaction_add(self, payload):
